Refactor GeoShapeIndexer by extracting polygon / line decomposers (#52422)
Refactor GeoShapeIndexer. We extract Polygon and Line decomposers which are in charge of breaking a shape around the dateline if needed.
